[Feature][Yvonne] Connected order-history to dashboard
